@@ -5,11 +5,11 @@ author: Simon Leier <simonleier@gmail.com>
 homepage: https://github.com/leisim/logger_flutter
 
 environment:
-  sdk: ">=2.2.2 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  logger: ">=0.7.0"
-  sensors: ^0.4.0+1
+  logger: "1.0.0"
+  sensors: ^2.0.3
